+++ core/AmRtpTransport.h
@@ -93,8 +93,8 @@ public:
     void addConnection(AmStreamConnection* conn);
     void removeConnection(AmStreamConnection* conn);
 
-    int send(AmRtpPacket* packet, AmStreamConnection::ConnectionType type);
-    int send(sockaddr_storage* raddr, unsigned char* buf, int size, AmStreamConnection::ConnectionType type);
+    ssize_t send(AmRtpPacket* packet, AmStreamConnection::ConnectionType type);
+    ssize_t send(sockaddr_storage* raddr, unsigned char* buf, int size, AmStreamConnection::ConnectionType type);
     int sendmsg(unsigned char* buf, int size);
 
     void allowStunConnection(sockaddr_storage* remote_addr, int priority);
@@ -142,18 +142,18 @@ public:
 protected:
     void addSrtpConnection(const string& remote_address, int remote_port, int srtp_ptrofile, const string& local_key, const string& remote_key);
     void addRtpConnection(const string& remote_address, int remote_port);
-    
-    int recv(int sd);
+
+    ssize_t recv(int sd);
     void recvPacket(int fd) override;
 
-    virtual void onPacket(unsigned char* buf, int size, sockaddr_storage& addr, struct timeval recvtime);
+    virtual void onPacket(unsigned char* buf, unsigned int size, sockaddr_storage& addr, struct timeval recvtime);
 
     void log_rcvd_packet(const char *buffer, int len, struct sockaddr_storage &recv_addr, AmStreamConnection::ConnectionType type);
     void log_sent_packet(const char *buffer, int len, struct sockaddr_storage &send_addr, AmStreamConnection::ConnectionType type);
 
     int getSrtpCredentialsBySdp(const SdpMedia& local_media, const SdpMedia& remote_media, string& local_key, string& remote_key);
     
-    AmStreamConnection::ConnectionType GetConnectionType(unsigned char* buf, int size);
+    AmStreamConnection::ConnectionType GetConnectionType(unsigned char* buf, unsigned int size);
     bool isStunMessage(unsigned char* buf, unsigned int size);
     bool isRTPMessage(unsigned char* buf, unsigned int size);
     bool isDTLSMessage(unsigned char* buf, unsigned int size);
